==> 00_extracting_dataset.ipynb <==
# Exctracting dataset for xG model

**ATTENTION:**

Notebook language: **R**

#install.packages("worldfootballR")

library(worldfootballR)

# Ligue 1
ligue1_2021_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2021)
ligue1_2020_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2020)
ligue1_2019_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2019)
ligue1_2018_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2018)
ligue1_2017_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2017)
ligue1_2016_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2016)
ligue1_2015_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2015)
ligue1_2014_shot_location <- understat_league_season_shots(league = "Ligue 1", season_start_year = 2014)

# Serie A
seriea_2021_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2021)
seriea_2020_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2020)
seriea_2019_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2019)
seriea_2018_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2018)
seriea_2017_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2017)
seriea_2016_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2016)
seriea_2015_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2015)
seriea_2014_shot_location <- understat_league_season_shots(league = "Serie A", season_start_year = 2014)

# Bundesliga
bundesliga_2021_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2021)
bundesliga_2020_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2020)
bundesliga_2019_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2019)
bundesliga_2018_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2018)
bundesliga_2017_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2017)
bundesliga_2016_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2016)
bundesliga_2015_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2015)
bundesliga_2014_shot_location <- understat_league_season_shots(league = "Bundesliga", season_start_year = 2014)

# La Liga
laliga_2021_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2021)
laliga_2020_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2020)
laliga_2019_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2019)
laliga_2018_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2018)
laliga_2017_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2017)
laliga_2016_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2016)
laliga_2015_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2015)
laliga_2014_shot_location <- understat_league_season_shots(league = "La liga", season_start_year = 2014)

# EPL
epl_2021_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2021)
epl_2020_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2020)
epl_2019_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2019)
epl_2018_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2018)
epl_2017_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2017)
epl_2016_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2016)
epl_2015_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2015)
epl_2014_shot_location <- understat_league_season_shots(league = "EPL", season_start_year = 2014)

# combining data
raw_data <- rbind(ligue1_2021_shot_location
                  ligue1_2020_shot_location,
                  ligue1_2019_shot_location,
                  ligue1_2018_shot_location,
                  ligue1_2017_shot_location,
                  ligue1_2016_shot_location,
                  ligue1_2015_shot_location,
                  ligue1_2014_shot_location,
                  
                  seriea_2021_shot_location,
                  seriea_2020_shot_location,
                  seriea_2019_shot_location,
                  seriea_2018_shot_location,
                  seriea_2017_shot_location,
                  seriea_2016_shot_location,
                  seriea_2015_shot_location,
                  seriea_2014_shot_location,
                  
                  bundesliga_2021_shot_location
                  bundesliga_2020_shot_location,
                  bundesliga_2019_shot_location,
                  bundesliga_2018_shot_location,
                  bundesliga_2017_shot_location,
                  bundesliga_2016_shot_location,
                  bundesliga_2015_shot_location,
                  bundesliga_2014_shot_location,
                  
                  laliga_2021_shot_location
                  laliga_2020_shot_location,
                  laliga_2019_shot_location,
                  laliga_2018_shot_location,
                  laliga_2017_shot_location,
                  laliga_2016_shot_location,
                  laliga_2015_shot_location,
                  laliga_2014_shot_location,
                  
                  epl_2021_shot_location
                  epl_2020_shot_location,
                  epl_2019_shot_location,
                  epl_2018_shot_location,
                  epl_2017_shot_location,
                  epl_2016_shot_location,
                  epl_2015_shot_location,
                  epl_2014_shot_location)

dir_name <- './data'

if (file.exists(dir_name)) {
 cat("The folder already exists")
} else {
 dir.create(dir_name)
}

write_csv(raw_data, './data/raw_data.csv')
