==> 02_preprocessing_exctracting_model.ipynb <==
# Exctracting model and dataset preprocessing

**ATTENTION:**

Notebook language: **R**

A model from a paper [EXPLAINABLE EXPECTED GOAL MODELS FOR PERFORMANCE
ANALYSIS IN FOOTBALL ANALYTICS](https://arxiv.org/pdf/2206.07212.pdf) will be used for further research. It is a `ranger` model trained on oversampled dataset.

## Preparing preprocessed dataset

df <- read.csv('./data/raw_data.csv')
df <- df[,-1]
head(df)

Remove `OwnGoal` - it is also a first preprocessing step 

library(dplyr)

df <- df %>% filter(result != "OwnGoal")
head(df)

Update `raw_data.csv` to avoid conflicts in future

write.csv(df, './data/raw_data.csv')

### Preprocessing

source('./scripts/preprocess.R') # preprocessing steps from the paper

df_preprocessed <- preprocess(df)
head(df_preprocessed)

df_preprocessed_1 <- df_preprocessed

factor_cols <- unlist(lapply(df_preprocessed_1, is.factor))
factor_cols <- names(factor_cols[factor_cols == TRUE])

levels_vector <- lapply(factor_cols, function(col){levels(df_preprocessed_1[,col])})
names(levels_vector) <- factor_cols
levels_vector

saveRDS(levels_vector, file = "./data/level_vector.RDS")

for (col in factor_cols){
    df_preprocessed_1[,col] <- as.integer(df_preprocessed_1[,col])
}

head(df_preprocessed_1)

write.csv(df_preprocessed_1, './data/data_preprocessed.csv')

## Check model prediction for a few first observations

library(ranger)

model <- readRDS('./model/model.RDS')

d <- df_preprocessed[1:11,]
d

predict(model, d)$predict

## Extracting model to `Python` format

library(reticulate)

py_save_object(model$forest, './model/model-imported.pickle')
