==> 03_loading_model.ipynb <==
# Loading model

**ATTENTION:**

Notebook language: **Python**

## Loading exctracted model data

import pickle5 as pickle

with open('./model/model-imported.pickle', 'rb') as fp:
    model = pickle.load(fp) 

model.keys()

## Converting attributes

model['num_trees'] = model.pop('num.trees')
model['child_node_ids'] = model.pop('child.nodeIDs')
model['split_var_ids'] = model.pop('split.varIDs')
model['split_values'] = model.pop('split.values')
model['is_ordered'] = model.pop('is.ordered')
model['class_values'] = model.pop('class.values')
model['terminal_class_counts'] = model.pop('terminal.class.counts')

model['num_trees'] = int(model['num_trees'])

def everything_as_int(x):
    if isinstance(x, int): return x
    elif isinstance(x, list): return [everything_as_int(y) for y in x]
    else: return int(x)

model["child_node_ids"] = everything_as_int(model["child_node_ids"])

## Loading data

import numpy as np
import pandas as pd

df = pd.read_csv('./data/data_preprocessed.csv', index_col=0)
df.head()

X = df.drop('status', axis = 1)
y = df['status']

## Loading `skranger` - `Python` interface to `ranger`

from scripts.RandomForestModified import RangerForestClassifierModified

rf = RangerForestClassifierModified(n_estimators = model['num_trees'], enable_tree_details = False)

`Fit` just adds parameters from already trained model; model is not trained here

rf.fit(X, y, {'forest': model, 'num_trees': model['num_trees']})

## Checking results

X.loc[1:11]

rf.predict(X.loc[1:11])

rf.predict_proba(X.loc[1:11])

np.array(y.loc[1:11])

## Save model

with open('./model/model.pickle', 'wb') as fp:
    pickle.dump(rf, fp) 

## Calculate metrics on the whole dataset

from scripts.calculate_metrics import calculate_metrics

calculate_metrics(rf, X, y)
