==> 05_aSHAP.ipynb <==
# Calculating aSHAP values (aggregated SHAP values)

**ATTENTION:**

Notebook language: **R**

SHAP values which were calculated in notebook ***04*** will be now transformed to be compatible with aSHAP implementation in *DALEX* package 

(in future *DALEX* can be used, now functions from a script will be called)

Theory:

As written on https://ema.drwhy.ai/breakDown.html#BDMethodGen Shapley values have the property called *local accuracy*:

$f(x_{i}) = v_{0} + \sum^{p}_{j=1} v(x_{i},j)$,

where: 

* $f$ - model
* $v_{0}$ - mean model prediction
* $x_{i}$ - i-th observation from a subset to be explained (let it be $X$)
* $v(x_{i},j)$ - SHAP value for i-th observation for j-th feature

Summing over the whole subset to be explained we get:

$\sum_{i=1}^{N} f(x_{i}) = \sum_{i=1}^{N} (v_{0} + \sum^{p}_{j=1} v(x_{i},j))$

$\frac{\sum_{i=1}^{N} f(x_{i})}{N} = \frac{\sum_{i=1}^{N} v_{0} + \sum_{i=1}^{N} \sum^{p}_{j=1} v(x_{i},j)}{N}$

$\overline{f(X)} = v_{0} + \sum^{p}_{j=1} \overline{v(X, j)}$

So, in plots below, *prediction* is an average of preditions over the set $X$, *intercept* is an average predictions over the training set and contributions are average contributions across set $X$.

source('./scripts/transform_shap.R')

source('./scripts/aSHAP.R')

## Function to transform and to create `shap_aggregated` objects

It is available in `./scripts/transform_shap.R`.

transform_all_tasks(data_dir = './data', 
                    results_dir = './results', 
                    label = 'ranger')

## Create plots

# create_aSHAP_plots_for_all_tasks('./results', max_features = 10, bg = 'white')
